Fixed warning

+++ src/Orbiter/ScreenHandler.cpp
@@ -1398,7 +1398,7 @@ bool ScreenHandler::MusicFullScreen_GridRendering(CallBackData *pData)
 					char *pImage = m_pOrbiter->ReadFileIntoBuffer(sName,size);
 					if( pImage )
 					{
-						m_pOrbiter->CMD_Update_Object_Image(TOSTRING(5551) ".0.0." TOSTRING(DESIGNOBJ_objCDCover_CONST),"jpg", pImage, size, "0");
+						m_pOrbiter->CMD_Update_Object_Image(TOSTRING(5551) ".0.0." TOSTRING(DESIGNOBJ_objCDCover_CONST),"jpg", pImage, static_cast<int>(size), "0");
 						bUpdatedPic=true;
 						delete[] pImage;
 					}
